+++ cpp/pmft/PMFTXYZ.cc
@@ -70,7 +70,7 @@ void PMFTXYZ::reduce()
  */
 void PMFTXYZ::accumulate(const locality::NeighborQuery* neighbor_query, quat<float>* query_orientations,
                          vec3<float>* query_points, unsigned int n_query_points,
-                         quat<float>* face_orientations, unsigned int n_faces,
+                         quat<float>* equiv_orientations, unsigned int num_equiv_orientations,
                          const locality::NeighborList* nlist, freud::locality::QueryArgs qargs)
 {
     // precalc some values for faster computation within the loop
@@ -78,20 +78,17 @@ void PMFTXYZ::accumulate(const locality::NeighborQuery* neighbor_query, quat<flo
     accumulateGeneral(neighbor_query, query_points, n_query_points, nlist, qargs,
                       [=](const freud::locality::NeighborBond& neighbor_bond) {
                           // create the reference point quaternion
-                          quat<float> ref_q(query_orientations[neighbor_bond.point_idx]);
+                          quat<float> ref_q(query_orientations[neighbor_bond.query_point_idx]);
                           // make sure that the particles are wrapped into the box
                           vec3<float> delta(bondVector(neighbor_bond, neighbor_query, query_points));
 
-                          for (unsigned int k = 0; k < n_faces; k++)
+                          for (unsigned int k = 0; k < num_equiv_orientations; k++)
                           {
-                              // create the extra quaternion
-                              quat<float> qe(face_orientations[util::ManagedArray<unsigned int>::getIndex(
-                                  {neighbor_query->getNPoints(), n_faces}, {neighbor_bond.point_idx, k})]);
                               // create point vector
                               vec3<float> v(delta);
                               // rotate the vector
                               v = rotate(conj(ref_q), v);
-                              v = rotate(qe, v);
+                              v = rotate(equiv_orientations[k], v);
 
                               m_local_histograms(v.x, v.y, v.z);
                           }