+++ cpp/pmft/PMFTXYZ.h
@@ -23,8 +23,8 @@ public:
         of the pcf
     */
     void accumulate(const locality::NeighborQuery* neighbor_query, quat<float>* query_orientations,
-                    vec3<float>* query_points, unsigned int n_query_points, quat<float>* face_orientations,
-                    unsigned int n_faces, const locality::NeighborList* nlist,
+                    vec3<float>* query_points, unsigned int n_query_points, quat<float>* equiv_orientations,
+                    unsigned int num_equiv_orientations, const locality::NeighborList* nlist,
                     freud::locality::QueryArgs qargs);
 
     //! \internal